# apps/miniapp/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Telegram Bot Manager API",
    "description": "REST API behind the Telegram Mini App admin panel.\n\nAuthenticate by posting the WebApp `initData` string to `/api/auth/telegram`,\nthen send the returned token as `Authorization: Bearer <token>` on every other\nrequest. Tokens are short-lived; renew them with `/api/auth/refresh` rather than\nre-posting `initData`, which Telegram only lets you use once.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/auth/telegram": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Exchange Telegram initData for an API session token",
        "operationId": "authenticateWithTelegram",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "401": {
            "description": "Session token missing, malformed or expired.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request failed validation.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/website": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Exchange a single-use bot-issued website token for a session",
        "operationId": "authenticateWithWebsiteToken",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebsiteLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "401": {
            "description": "Session token missing, malformed or expired.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request failed validation.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Renew a session token that has not expired yet",
        "operationId": "refreshSession",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "401": {
            "description": "Session token missing, malformed or expired.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request failed validation.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "TelegramSession": []
          }
        ]
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "The caller behind the current session token",
        "operationId": "getCurrentUser",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthUser"
                }
              }
            }
          },
          "401": {
            "description": "Session token missing, malformed or expired.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request failed validation.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "TelegramSession": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthRequest": {
        "properties": {
          "init_data": {
            "type": "string",
            "maxLength": 8192,
            "minLength": 1,
            "title": "Init Data"
          }
        },
        "type": "object",
        "required": [
          "init_data"
        ],
        "title": "AuthRequest",
        "description": "Raw `initData` string handed over by the Telegram WebApp SDK."
      },
      "AuthResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In"
          },
          "user": {
            "$ref": "#/components/schemas/AuthUser"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "expires_in",
          "user"
        ],
        "title": "AuthResponse"
      },
      "AuthUser": {
        "properties": {
          "tg_user_id": {
            "type": "integer",
            "title": "Tg User Id"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          },
          "first_name": {
            "type": "string",
            "title": "First Name",
            "default": ""
          },
          "last_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Name"
          },
          "language_code": {
            "type": "string",
            "title": "Language Code",
            "default": "en"
          },
          "is_superadmin": {
            "type": "boolean",
            "title": "Is Superadmin",
            "default": false
          },
          "is_premium": {
            "type": "boolean",
            "title": "Is Premium",
            "default": false
          },
          "has_photo": {
            "type": "boolean",
            "title": "Has Photo",
            "default": false
          },
          "photo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Url"
          }
        },
        "type": "object",
        "required": [
          "tg_user_id"
        ],
        "title": "AuthUser"
      },
...
      "Problem": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "default": "about:blank"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "integer",
            "title": "Status"
          },
          "detail": {
            "type": "string",
            "title": "Detail",
            "default": ""
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance"
          },
          "code": {
            "type": "string",
            "title": "Code",
            "default": "domain-error"
          },
          "context": {
            "additionalProperties": true,
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "title",
          "status"
        ],
        "title": "Problem",
        "description": "RFC7807-like error body returned by the API exception handler.\n\n`title` is localized for the caller and safe to show; `detail` is the\ndeveloper-facing message. `code` is the stable machine-readable discriminator\nthe Mini App branches on, and `context` carries whatever the error knew (the\nrequired plan for a paywall, the offending field for a validation failure)."
      },
...
      "WebsiteLoginRequest": {
        "properties": {
          "token": {
            "type": "string",
            "maxLength": 256,
            "minLength": 32,
            "title": "Token"
          }
        },
        "type": "object",
        "required": [
          "token"
        ],
        "title": "WebsiteLoginRequest"
      }
    },
    "securitySchemes": {
      "TelegramSession": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}